refactor(smoke): split pi smoke checks into subsystem modules
