+++ src/yoyopod/cli/pi/smoke/__init__.py
@@ -1,0 +1,38 @@
+"""Smoke test helpers and command entrypoint package."""
+
+from __future__ import annotations
+
+from .command import (
+    _load_app_config,
+    _load_media_config,
+    smoke,
+    smoke_app,
+)
+from .display import _display_check
+from .environment import _environment_check
+from .input import _input_check
+from .lvgl import _lvgl_soak_check
+from .music import _music_check, _prepare_music_validation_library
+from .power import _power_check
+from .report import _print_summary
+from .rtc import _rtc_check
+from .types import CheckResult
+from .voip import _voip_check
+
+__all__ = [
+    "CheckResult",
+    "_environment_check",
+    "_display_check",
+    "_input_check",
+    "_power_check",
+    "_rtc_check",
+    "_prepare_music_validation_library",
+    "_music_check",
+    "_voip_check",
+    "_lvgl_soak_check",
+    "_print_summary",
+    "_load_app_config",
+    "_load_media_config",
+    "smoke",
+    "smoke_app",
+]
