+++ src/yoyopod/cli/pi/smoke/environment.py
@@ -1,0 +1,25 @@
+"""Environment-level smoke check helpers."""
+
+from __future__ import annotations
+
+import platform
+
+from .types import CheckResult
+
+
+def _environment_check() -> CheckResult:
+    """Capture the current execution environment."""
+    system = platform.system()
+    machine = platform.machine()
+    python_version = platform.python_version()
+
+    if system == "Linux" and ("arm" in machine.lower() or "aarch" in machine.lower()):
+        status = "pass"
+    else:
+        status = "warn"
+
+    return CheckResult(
+        name="environment",
+        status=status,
+        details=f"system={system}, machine={machine}, python={python_version}",
+    )
